+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import logo from '~/assets/icons/mutiny-logo.svg';
+import { SafeArea } from "~/components/layout";
 import BalanceBox from "~/components/BalanceBox";
-import SafeArea from "~/components/SafeArea";
 import NavBar from "~/components/NavBar";
 
 // TODO: use this reload prompt for real
